+++ AKZ.API/Models/ProcessingRecord.cs
@@ -20,47 +20,47 @@ public class ProcessingRecord : BaseEntity
     [Required]
     public string WorkerNames { get; set; } = string.Empty;
 
-    public int Count { get; set; }
-    public int RemainingCount { get; set; }
+    public double Count { get; set; }
+    public double RemainingCount { get; set; }
 
     [Column(TypeName = "decimal(18,4)")]
     public decimal UnitWeight { get; set; }
 
     [Column(TypeName = "decimal(18,4)")]
     public decimal RedWeight { get; set; }
-    public int RedCount { get; set; }
+    public double RedCount { get; set; }
 
     [Column(TypeName = "decimal(18,4)")]
     public decimal WhiteWeight { get; set; }
-    public int WhiteCount { get; set; }
+    public double WhiteCount { get; set; }
 
     [Column(TypeName = "decimal(18,4)")]
     public decimal SpecialWeight { get; set; }
-    public int SpecialCount { get; set; }
+    public double SpecialCount { get; set; }
 
     [Column(TypeName = "decimal(18,4)")]
     public decimal NaturalWeight { get; set; }
-    public int NaturalCount { get; set; }
+    public double NaturalCount { get; set; }
 
     [Column(TypeName = "decimal(18,4)")]
     public decimal NaturalWhiteWeight { get; set; }
-    public int NaturalWhiteCount { get; set; }
+    public double NaturalWhiteCount { get; set; }
 
     [Column(TypeName = "decimal(18,4)")]
     public decimal NaturalRedWeight { get; set; }
-    public int NaturalRedCount { get; set; }
+    public double NaturalRedCount { get; set; }
 
     [Column(TypeName = "decimal(18,4)")]
     public decimal ShortCutWeight { get; set; }
-    public int ShortCutCount { get; set; }
+    public double ShortCutCount { get; set; }
 
     [Column(TypeName = "decimal(18,4)")]
     public decimal ArtificialWeight { get; set; }
-    public int ArtificialCount { get; set; }
+    public double ArtificialCount { get; set; }
 
     [Column(TypeName = "decimal(18,4)")]
     public decimal ShortWeight { get; set; }
-    public int ShortCount { get; set; }
+    public double ShortCount { get; set; }
 
     [Column(TypeName = "decimal(18,4)")]
     public decimal LossWeight { get; set; }
@@ -78,15 +78,15 @@ public class ProcessingRecord : BaseEntity
     public decimal Difference { get; set; }
 
     // Remaining counts per category
-    public int RemRedCount { get; set; }
-    public int RemWhiteCount { get; set; }
-    public int RemSpecialCount { get; set; }
-    public int RemNaturalCount { get; set; }
-    public int RemNaturalWhiteCount { get; set; }
-    public int RemNaturalRedCount { get; set; }
-    public int RemShortCutCount { get; set; }
-    public int RemArtificialCount { get; set; }
-    public int RemShortCount { get; set; }
+    public double RemRedCount { get; set; }
+    public double RemWhiteCount { get; set; }
+    public double RemSpecialCount { get; set; }
+    public double RemNaturalCount { get; set; }
+    public double RemNaturalWhiteCount { get; set; }
+    public double RemNaturalRedCount { get; set; }
+    public double RemShortCutCount { get; set; }
+    public double RemArtificialCount { get; set; }
+    public double RemShortCount { get; set; }
 
     // Remaining weights per category
     [Column(TypeName = "decimal(18,4)")]
